requirements.txt
pandas
numpy
seaborn
pymongo[srv]
certifi
python-dotenv
pyyaml
scikit-learn==1.1.3
imblearn==0.0
dill
xgboost
azure-storage-blob
-e .